feat: change room picture and name

+++ src/screens/EditProfile/EditProfile.tsx
@@ -1,25 +1,36 @@
 import React, { FC, useEffect, useState } from 'react';
-import { Alert, BackHandler, View } from 'react-native';
+import { Alert, BackHandler, StatusBar, StyleSheet, View } from 'react-native';
 
+import DocumentPicker from 'react-native-document-picker';
 import FastImage from 'react-native-fast-image';
 import FileSystem from 'react-native-fs';
 import { ScrollView, TouchableOpacity } from 'react-native-gesture-handler';
-import { Avatar, Button, FAB as Fab, HelperText, TextInput } from 'react-native-paper';
+import {
+  ActivityIndicator,
+  Appbar,
+  Avatar,
+  Button,
+  FAB as Fab,
+  HelperText,
+  TextInput,
+} from 'react-native-paper';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { useDatabase } from '@nozbe/watermelondb/hooks';
 import CameraRoll from '@react-native-community/cameraroll';
 
 import { USER_NAME_MAX_LENGTH } from '!/config';
-import { useBackHandlerOnFocus } from '!/hooks/use-back-handler';
 import useDimensions from '!/hooks/use-dimensions';
+import useFocusEffect from '!/hooks/use-focus-effect';
 import useInput from '!/hooks/use-input';
+import useMethod from '!/hooks/use-method';
 import usePress from '!/hooks/use-press';
 import useTheme from '!/hooks/use-theme';
 import useTranslation from '!/hooks/use-translation';
 import { userUpdater } from '!/models/UserModel';
 import { uploadMedia } from '!/services/remote-media';
 import { useStores } from '!/stores';
-import { MainNavigationProp, MainRouteProp, StackHeaderRightProps } from '!/types';
+import { HeaderOptions, MainNavigationProp, MainRouteProp, StackHeaderRightProps } from '!/types';
+import getStatusBarColor from '!/utils/get-status-bar-color';
 
 import styles from './styles';
 
@@ -32,40 +43,66 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
   const database = useDatabase();
   const [winWidth] = useDimensions('window');
   const { authStore, syncStore } = useStores();
-  const { colors, dark, grid, gridBigger } = useTheme();
+  const { colors, dark, mode, grid, gridBigger } = useTheme();
   const { t } = useTranslation();
 
   const { params } = route;
 
+  const [pictureUri, setPictureUri] = useState('');
+
   const [isUploadingPic, setIsUploadingPic] = useState(false);
-  const [nameError, setNameError] = useState('');
+  const [errorMessage, setErrorMessage] = useState('');
 
-  const nameInput = useInput(authStore.user.name, () => setNameError(''));
+  const nameInput = useInput(authStore.user.name, () => setErrorMessage(''));
 
-  const handleChangePicture = usePress(() => {
+  const handlePressBack = usePress(() => {
+    navigation.navigate('Settings', { reload: true });
+  });
+
+  const handleOpenAttachmentPicker = usePress(() => {
     requestAnimationFrame(() => {
-      navigation.navigate('Camera', {
-        screenNameAfterPicture: 'EditProfile',
-        disableRecordVideo: true,
-        initialCameraType: 'front',
-        showCameraMask: true,
+      navigation.navigate('AttachmentPickerModal', {
+        callbackScreen: 'EditProfile',
+        types: ['camera', 'image'],
       });
     });
   });
 
+  const handlePickImage = useMethod(async () => {
+    try {
+      const fileChosen = await DocumentPicker.pick({
+        type: [DocumentPicker.types.images],
+      });
+
+      navigation.setParams({ picturesTaken: undefined });
+      requestAnimationFrame(() => {
+        setPictureUri(fileChosen.uri);
+      });
+    } catch (err) {
+      if (DocumentPicker.isCancel(err)) {
+        // User cancelled the picker, exit any dialogs or menus and move on
+      } else {
+        console.error(err);
+      }
+    }
+  });
+
   const handleDiscardPicture = usePress(async () => {
+    navigation.setParams({ picturesTaken: undefined });
     requestAnimationFrame(() => {
-      navigation.setParams({ ...params, picturesTaken: undefined });
+      setPictureUri('');
     });
 
-    try {
-      const file = params!.picturesTaken![0].uri!;
-      const fileInfo = await FileSystem.stat(file);
+    if (params?.picturesTaken?.[0].uri) {
+      try {
+        const file = params.picturesTaken[0].uri;
+        const fileInfo = await FileSystem.stat(file);
 
-      await CameraRoll.deletePhotos([fileInfo.originalFilepath]);
-      await FileSystem.unlink(fileInfo.originalFilepath);
-    } catch (err) {
-      console.log(err);
+        await CameraRoll.deletePhotos([fileInfo.originalFilepath]);
+        await FileSystem.unlink(fileInfo.originalFilepath);
+      } catch (err) {
+        console.log(err);
+      }
     }
   });
 
@@ -79,7 +116,7 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
           text: t('label.yes'),
           onPress: () => {
             requestAnimationFrame(() => {
-              navigation.setParams({ ...params, picturesTaken: undefined });
+              navigation.setParams({ picturesTaken: undefined });
             });
 
             void database.action(async () => {
@@ -101,7 +138,7 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
       const name = nameInput.value.trim();
 
       if (!name) {
-        setNameError(t('helper.fieldRequired'));
+        setErrorMessage(t('helper.fieldRequired'));
         return;
       }
 
@@ -109,9 +146,9 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
 
       let remoteUri: string | undefined;
 
-      // Upload picture
-      if (params?.picturesTaken?.[0]) {
-        const res = await uploadMedia(params.picturesTaken[0].uri!);
+      const pictureUriToUpload = params?.picturesTaken?.[0].uri ?? pictureUri;
+      if (pictureUriToUpload) {
+        const res = await uploadMedia(pictureUriToUpload);
         remoteUri = res.secure_url;
       }
 
@@ -135,12 +172,17 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
       }
 
       requestAnimationFrame(() => {
-        navigation.reset({
-          index: 0,
-          routes: [{ name: 'Home' }],
-        });
+        if (params?.isEditing) {
+          handlePressBack();
+        } else {
+          navigation.reset({
+            index: 0,
+            routes: [{ name: 'Home' }],
+          });
+        }
       });
-    } catch (e) {
+    } catch (err) {
+      console.log(err);
       Alert.alert(t('title.oops'));
       setIsUploadingPic(false);
     }
@@ -155,33 +197,82 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
     });
   });
 
-  useBackHandlerOnFocus(() => {
-    if (params?.isEditing) {
-      return false;
+  useFocusEffect(() => {
+    const type = `${params?.attachmentType || ''}`;
+
+    navigation.setParams({ attachmentType: undefined });
+
+    if (type === 'camera') {
+      requestAnimationFrame(() => {
+        setPictureUri('');
+        navigation.navigate('Camera', {
+          screenNameAfterPicture: 'EditProfile',
+          disableRecordVideo: true,
+          initialCameraType: 'front',
+          showCameraMask: true,
+        });
+      });
+    } else if (type === 'image') {
+      void handlePickImage();
     }
-    Alert.alert(t('title.closeTheApp'), t('alert.youCanAlwaysChangeYourProfileLater'), [
-      { text: t('label.no') },
-      { text: t('label.yes'), onPress: () => BackHandler.exitApp() },
-    ]);
-    return true;
-  }, [params?.isEditing, t]);
+
+    const backHandler = BackHandler.addEventListener('hardwareBackPress', () => {
+      if (params?.isEditing) {
+        handlePressBack();
+        return true;
+      }
+      Alert.alert(t('title.closeTheApp'), t('alert.youCanAlwaysChangeYourProfileLater'), [
+        { text: t('label.no') },
+        { text: t('label.yes'), onPress: () => BackHandler.exitApp() },
+      ]);
+      return true;
+    });
+
+    return () => {
+      backHandler.remove();
+    };
+  }, [handlePickImage, handlePressBack, navigation, params?.attachmentType, params?.isEditing, t]);
 
   useEffect(() => {
+    const elevation = 4;
+    const bgColor = getStatusBarColor(elevation, colors, dark, mode);
+
+    StatusBar.setHidden(false);
+    StatusBar.setBackgroundColor(bgColor, true);
+    StatusBar.setBarStyle(colors.statusBarText);
+    StatusBar.setTranslucent(false);
+
     navigation.setOptions({
       headerShown: params?.isEditing,
-      headerRight: !params?.isEditing
-        ? ({ tintColor }: StackHeaderRightProps) => (
-            <Button
-              color={dark ? colors.primary : tintColor}
-              disabled={isUploadingPic}
-              onPress={handleSignOut}
-            >
-              {t('signOut')}
-            </Button>
-          )
-        : undefined,
-    });
-  }, [colors.primary, dark, handleSignOut, isUploadingPic, navigation, params?.isEditing, t]);
+      handlePressBack: params?.isEditing ? handlePressBack : undefined,
+      headerRight: ({ tintColor }: StackHeaderRightProps) =>
+        !params?.isEditing ? (
+          <Button color={dark ? colors.primary : tintColor} disabled={isUploadingPic} onPress={handleSignOut}>
+            {t('signOut')}
+          </Button>
+        ) : (
+          <Appbar.Action
+            color={tintColor}
+            disabled={isUploadingPic}
+            icon='check-bold'
+            onPress={handleFinishSetup}
+          />
+        ),
+    } as HeaderOptions);
+  }, [
+    colors,
+    dark,
+    handleFinishSetup,
+    handlePressBack,
+    handleSignOut,
+    isUploadingPic,
+    mode,
+    navigation,
+    params?.isEditing,
+    t,
+  ]);
+
+  const avatarSize = Math.min(200, winWidth - grid * 10);
 
   return (
     <ScrollView contentContainerStyle={styles.contentContainer}>
@@ -192,19 +283,26 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
               delayLongPress={600}
               disabled={isUploadingPic}
               onLongPress={handleRemoveSavedPicture}
-              onPress={handleChangePicture}
+              onPress={handleOpenAttachmentPicker}
             >
-              {params?.picturesTaken?.[0] ? (
+              {pictureUri ? (
                 <Avatar.Image
                   ImageComponent={FastImage}
-                  size={Math.min(200, winWidth - grid * 10)}
+                  size={avatarSize}
+                  source={{ uri: pictureUri }}
+                  style={styles.avatar}
+                />
+              ) : params?.picturesTaken?.[0] ? (
+                <Avatar.Image
+                  ImageComponent={FastImage}
+                  size={avatarSize}
                   source={{ uri: params.picturesTaken[0].uri }}
                   style={styles.avatar}
                 />
               ) : authStore.user.pictureUri ? (
                 <Avatar.Image
                   ImageComponent={FastImage}
-                  size={Math.min(200, winWidth - grid * 10)}
+                  size={avatarSize}
                   source={{ uri: authStore.user.pictureUri }}
                   style={styles.avatar}
                 />
@@ -212,13 +310,13 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
                 <Avatar.Icon
                   color={colors.textOnPrimary}
                   icon='image-plus'
-                  size={Math.min(200, winWidth - grid * 10)}
+                  size={avatarSize}
                   style={styles.avatar}
                 />
               )}
             </TouchableOpacity>
 
-            {params?.picturesTaken?.[0] ? (
+            {pictureUri ?? params?.picturesTaken?.[0] ? (
               <Fab
                 color={colors.textOnAccent}
                 disabled={isUploadingPic}
@@ -227,6 +325,10 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
                 small
                 style={styles.fabDiscardPicture}
               />
+            ) : null}
+
+            {isUploadingPic ? (
+              <ActivityIndicator color={colors.accent} size='large' style={StyleSheet.absoluteFill} />
             ) : null}
           </View>
           <HelperText
@@ -241,28 +343,31 @@ const EditProfile: FC<Props> = ({ navigation, route }) => {
             autoCapitalize='sentences'
             blurOnSubmit
             disabled={isUploadingPic}
-            error={!!nameError}
+            error={!!errorMessage}
             label={t('label.userName')}
             maxLength={USER_NAME_MAX_LENGTH}
             mode='outlined'
             returnKeyType='done'
             {...nameInput}
           />
-          <HelperText type='error' visible>
-            {nameError || null}
+          <HelperText type={errorMessage ? 'error' : 'info'} visible>
+            {errorMessage ||
+              t('helper.charactersLeft', { count: USER_NAME_MAX_LENGTH - nameInput.value.trim().length })}
           </HelperText>
         </View>
 
-        <Button
-          disabled={isUploadingPic}
-          labelStyle={{ color: isUploadingPic ? colors.disabled : colors.textOnPrimary }}
-          loading={isUploadingPic}
-          mode='contained'
-          onPress={handleFinishSetup}
-          style={styles.button}
-        >
-          {t(params?.isEditing ? 'label.save' : 'label.continue')}
-        </Button>
+        {!params?.isEditing ? (
+          <Button
+            disabled={isUploadingPic}
+            labelStyle={{ color: isUploadingPic ? colors.disabled : colors.textOnPrimary }}
+            loading={isUploadingPic}
+            mode='contained'
+            onPress={handleFinishSetup}
+            style={styles.button}
+          >
+            {t('label.continue')}
+          </Button>
+        ) : null}
       </SafeAreaView>
     </ScrollView>
   );
